backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
